WIP @params

+++ protocol/synthetix/contracts/interfaces/IVaultModule.sol
@@ -9,18 +9,29 @@ pragma solidity ^0.8.0;
 interface IVaultModule {
     /**
      * @notice Emitted when {sender} updates the delegation of collateral in the specified staking position.
+     * @param accountId The id of the account whose position was updated.
+     * @param poolId The id of the pool in which the position was updated.
+     * @param collateralType The address of the collateral associated to the position.
+     * @param amount The new amount of the position.
+     * @param leverage The new leverage value of the position.
+     * @param sender The address that triggered the update of the position.
      */
     event DelegationUpdated(
         uint128 indexed accountId,
         uint128 indexed poolId,
         address collateralType,
-        uint amount,
-        uint leverage,
+        uint256 amount,
+        uint256 leverage,
         address indexed sender
     );
 
     /**
      * @notice Updates an account's delegated collateral amount for the specified pool and collateral type pair.
+     * @param accountId The id of the account associated with the position that will be updated.
+     * @param poolId The id of the pool associated with the position.
+     * @param collateralType The address of the collateral used in the position.
+     * @param amount The new amount of collateral delegated in the position.
+     * @param leverage The new leverage amount used in the position.
      *
      * Requirements:
      *
@@ -34,46 +45,65 @@ interface IVaultModule {
         uint128 accountId,
         uint128 poolId,
         address collateralType,
-        uint amount,
-        uint leverage
+        uint256 amount,
+        uint256 leverage
     ) external;
 
     /**
      * @notice Returns the collateralization ratio of the specified staking position. If debt is negative, this function will return 0.
      * @dev Call this function using `callStatic` to treat it as a view function.
      * @dev The return value is a percentage with 18 decimals places.
+     * @param accountId The id of the account whose collateralization ratio is being queried.
+     * @param poolId The id of the pool in which the account's position is held.
+     * @param collateralType The address of the collateral used in the queried position.
+     * @returns The collateralization ratio of the position (collateral / debt).
      */
     function getPositionCollateralizationRatio(
         uint128 accountId,
         uint128 poolId,
         address collateralType
-    ) external returns (uint);
+    ) external returns (uint256);
 
     /**
      * @notice Returns the debt of the specified staking position. Credit is expressed as negative debt.
      * @dev This is not a view function, and actually updates the entire debt distribution chain.
      * @dev To call this externally as a view function, use `staticall`.
+     * @param accountId The id of the account being queried.
+     * @param poolId The id of the pool in which the account's position is held.
+     * @param collateralType The address of the collateral used in the queried position.
+     * @returns The amount of USD debt held by the position.
      */
     function getPositionDebt(
         uint128 accountId,
         uint128 poolId,
         address collateralType
-    ) external returns (int);
+    ) external returns (int256);
 
     /**
      * @notice Returns the amount and value of the collateral associated with the specified staking position.
      * @dev Call this function using `callStatic` to treat it as a view function.
      * @dev collateralAmount is represented as an integer with 18 decimals.
      * @dev collateralValue is represented as an integer with the number of decimals specified by the collateralType.
+     * @param accountId The id of the account being queried.
+     * @param poolId The id of the pool in which the account's position is held.
+     * @param collateralType The address of the collateral used in the queried position.
+     * @returns The amount of collateral used in the position.
      */
     function getPositionCollateral(
         uint128 accountId,
         uint128 poolId,
         address collateralType
-    ) external view returns (uint collateralAmount, uint collateralValue);
+    ) external view returns (uint256 collateralAmount, uint256 collateralValue);
 
     /**
      * @notice Returns all information pertaining to a specified staking position in the vault module.
+     * @param accountId The id of the account being queried.
+     * @param poolId The id of the pool in which the account's position is held.
+     * @param collateralType The address of the collateral used in the queried position.
+     * @returns collateralAmount The amount of collateral used in the position.
+     * @returns collateralValue The value of the collateral used in the position.
+     * @returns debt The amount of USD debt held in the position.
+     * @returns collateralizationRatio The collateralization ratio of the position (collateral / debt).
      **/
     function getPosition(
         uint128 accountId,
@@ -82,37 +112,44 @@ interface IVaultModule {
     )
         external
         returns (
-            uint collateralAmount,
-            uint collateralValue,
-            int debt,
-            uint collateralizationRatio
+            uint256 collateralAmount,
+            uint256 collateralValue,
+            int256 debt,
+            uint256 collateralizationRatio
         );
 
     /**
      * @notice Returns the total debt (or credit) that the vault is responsible for. Credit is expressed as negative debt.
      * @dev This is not a view function, and actually updates the entire debt distribution chain.
      * @dev To call this externally as a view function, use `staticall`.
+     * @param poolId The id of the pool that owns the vault whose debt is being queried.
+     * @param collateralType The address of the collateral of the associated vault.
+     * @returns The overall debt of the vault.
      **/
-    function getVaultDebt(uint128 poolId, address collateralType) external returns (int);
+    function getVaultDebt(uint128 poolId, address collateralType) external returns (int256);
 
     /**
      * @notice Returns the amount and value of the collateral held by the vault.
      * @dev Call this function using `callStatic` to treat it as a view function.
      * @dev collateralAmount is represented as an integer with 18 decimals.
      * @dev collateralValue is represented as an integer with the number of decimals specified by the collateralType.
+     * @param poolId The id of the pool that owns the vault whose collateral is being queried.
+     * @param collateralType The address of the collateral of the associated vault.
+     * @returns The collateral of the vault.
      */
-    function getVaultCollateral(
-        uint128 poolId,
-        address collateralType
-    ) external returns (uint collateralAmount, uint collateralValue);
+    function getVaultCollateral(uint128 poolId, address collateralType)
+        external
+        returns (uint256 collateralAmount, uint256 collateralValue);
 
     /**
      * @notice Returns the collateralization ratio of the vault. If debt is negative, this function will return 0.
      * @dev Call this function using `callStatic` to treat it as a view function.
      * @dev The return value is a percentage with 18 decimals places.
+     * @param poolId The id of the pool that owns the vault whose collateralization ratio is being queried.
+     * @param collateralType The address of the collateral of the associated vault.
+     * @returns The collateralization ratio of the vault.
      */
-    function getVaultCollateralRatio(
-        uint128 poolId,
-        address collateralType
-    ) external returns (uint);
+    function getVaultCollateralRatio(uint128 poolId, address collateralType)
+        external
+        returns (uint256);
 }
